+++ eslint.config.mjs
@@ -55,6 +55,7 @@ export default tseslint.config(
       ],
       'no-alert': 'warn',
       'no-debugger': 'warn',
+      'react/no-children-prop': 'off',
     },
     linterOptions: {
       reportUnusedDisableDirectives: 'off',
