+++ packages/cmsui/package.json
@@ -54,6 +54,7 @@
     "@plone/blocks": "workspace:*",
     "@plone/client": "workspace:*",
     "@plone/components": "workspace:*",
+    "@plone/helpers": "workspace:*",
     "@plone/react-router": "workspace:*",
     "@plone/registry": "workspace:*",
     "@tanstack/react-form": "^1.3.3",
