+++ packages/helpers/src/atoms.ts
@@ -1,0 +1,18 @@
+import { useHydrateAtoms } from 'jotai/utils';
+import type { WritableAtom } from 'jotai';
+import type { ReactNode } from 'react';
+
+export const InitAtoms = ({
+  atomValues,
+  children,
+}: {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  atomValues: Iterable<
+    readonly [WritableAtom<unknown, [any], unknown>, unknown]
+  >;
+  children: ReactNode;
+}) => {
+  // initialising on state with prop on render here
+  useHydrateAtoms(new Map(atomValues));
+  return children;
+};
